==> 3-RScripts/0-packages.R <==
#these are all the packages I used in my analysis
#some (many) of these are not necessary for what I ended up including in my report



library(dplyr) 
library(forcats)
library(ggplot2)
library(esquisse)
library(mosaic)
library(tidyr)
library(tidymodels)
library(lme4)
library(lmerTest)

library(visreg)
library(dotwhisker)
library(performance)

library(osmdata)
library(sf) 
library(mapview)
library(ggrepel)

library(merDeriv)
library(marginaleffects)
library(lubridate)

library(sjPlot)
library(sjmisc)
library(sjlabelled)
