==> 3-RScripts/7-citations.R <==
source("3-RScripts/packages.R")

# citations for model computation
citation(package = "lme4")
citation(package = "lmerTest")

#citations for model visualization 
citation(package = "ggplot2")
citation(package = "marginaleffects")
citation(package = "lubridate")

#citations for model table creation
library(sjPlot)
library(sjmisc)
library(sjlabelled)
